+++ src/_utils.js
@@ -1,0 +1,6 @@
+export const _setTimeout = (callback = () => {}, time = 0) => {
+  const timer = window.setTimeout(() => {
+    callback();
+    window.clearTimeout(timer);
+  }, time)
+};