+++ src/tools/index.js
@@ -10,6 +10,8 @@ import storageWorker from './storageWorker';
 import useDebounce from './useDebounce';
 import clipboard from './clipboard';
 import typeOf from './typeOf';
+import moneyToChinese from './moneyToChinese';
+import identityCard from './identityCard';
 
 export default {
   delay,
@@ -24,4 +26,6 @@ export default {
   objectToString,
   getUrlParameter,
   genNonDuplicateID,
+  moneyToChinese,
+  identityCard,
 };